+++ src/main/java/net/countercraft/movecraft/utils/WorldguardUtils.java
@@ -1,0 +1,39 @@
+package net.countercraft.movecraft.utils;
+
+import com.sk89q.worldguard.protection.regions.ProtectedRegion;
+import org.bukkit.Bukkit;
+import org.bukkit.OfflinePlayer;
+
+import java.util.UUID;
+
+public class WorldguardUtils {
+    public static String getRegionOwnerList(ProtectedRegion tRegion) {
+        StringBuilder output = new StringBuilder();
+        if (tRegion == null)
+            return output.toString();
+        boolean first = true;
+        if (tRegion.getOwners().getUniqueIds().size() > 0) {
+            for (UUID uid : tRegion.getOwners().getUniqueIds()) {
+                if (!first)
+                    output.append(",");
+                else
+                    first = false;
+                OfflinePlayer offP = Bukkit.getOfflinePlayer(uid);
+                if (offP.getName() == null)
+                    output.append(uid.toString());
+                else
+                    output.append(offP.getName());
+            }
+        }
+        if (tRegion.getOwners().getPlayers().size() > 0) {
+            for (String player : tRegion.getOwners().getPlayers()) {
+                if (!first)
+                    output.append(",");
+                else
+                    first = false;
+                output.append(player);
+            }
+        }
+        return output.toString();
+    }
+}
